requirements.txt
trimesh
pymeshlab==2022.2.post3
pygltflib
xatlas
open3d
omegaconf
pyyaml
configargparse
transformers
diffusers
accelerate
pytorch-lightning
opencv-python
huggingface-hub
safetensors
scikit-image
pybind11
timm

meshlib
